Make DPoPStore implementation optional
