+++ docs/oidc-client-ts.api.md
@@ -306,12 +306,14 @@ export class OidcClient {
     createSigninRequest({ state, request, request_uri, request_type, id_token_hint, login_hint, skipUserInfo, nonce, url_state, response_type, scope, redirect_uri, prompt, display, max_age, ui_locales, acr_values, resource, response_mode, extraQueryParams, extraTokenParams, }: CreateSigninRequestArgs): Promise<SigninRequest>;
     // (undocumented)
     createSignoutRequest({ state, id_token_hint, client_id, request_type, post_logout_redirect_uri, extraQueryParams, }?: CreateSignoutRequestArgs): Promise<SignoutRequest>;
+    // Warning: (ae-forgotten-export) The symbol "IndexDbDPoPStore" needs to be exported by the entry point index.d.ts
+    //
+    // (undocumented)
+    protected readonly _dpopStore: IndexDbDPoPStore | undefined;
     // Warning: (ae-forgotten-export) The symbol "DPoPStore" needs to be exported by the entry point index.d.ts
     //
     // (undocumented)
-    protected readonly _dpopStore: DPoPStore | undefined;
-    // (undocumented)
-    getDpopProof(dpopStore: DPoPStore): Promise<string>;
+    getDpopProof(dpopStore: DPoPStore<CryptoKeyPair>): Promise<string>;
     // (undocumented)
     protected readonly _logger: Logger;
     // (undocumented)
@@ -359,6 +361,7 @@ export interface OidcClientSettings {
     disablePKCE?: boolean;
     display?: string;
     dpop?: boolean;
+    dpopStore?: DPoPStore<CryptoKeyPair>;
     extraHeaders?: Record<string, ExtraHeader>;
     extraQueryParams?: Record<string, string | number | boolean>;
     // (undocumented)
@@ -391,7 +394,7 @@ export interface OidcClientSettings {
 
 // @public
 export class OidcClientSettingsStore {
-    constructor({ authority, metadataUrl, metadata, signingKeys, metadataSeed, client_id, client_secret, response_type, scope, redirect_uri, post_logout_redirect_uri, client_authentication, prompt, display, max_age, ui_locales, acr_values, resource, response_mode, filterProtocolClaims, loadUserInfo, staleStateAgeInSeconds, mergeClaimsStrategy, disablePKCE, stateStore, revokeTokenAdditionalContentTypes, fetchRequestCredentials, refreshTokenAllowedScope, extraQueryParams, extraTokenParams, extraHeaders, dpop, }: OidcClientSettings);
+    constructor({ authority, metadataUrl, metadata, signingKeys, metadataSeed, client_id, client_secret, response_type, scope, redirect_uri, post_logout_redirect_uri, client_authentication, prompt, display, max_age, ui_locales, acr_values, resource, response_mode, filterProtocolClaims, loadUserInfo, staleStateAgeInSeconds, mergeClaimsStrategy, disablePKCE, stateStore, revokeTokenAdditionalContentTypes, fetchRequestCredentials, refreshTokenAllowedScope, extraQueryParams, extraTokenParams, extraHeaders, dpop, dpopStore, }: OidcClientSettings);
     // (undocumented)
     readonly acr_values: string | undefined;
     // (undocumented)
@@ -408,6 +411,8 @@ export class OidcClientSettingsStore {
     readonly display: string | undefined;
     // (undocumented)
     readonly dpop: boolean | undefined;
+    // (undocumented)
+    readonly dpopStore: DPoPStore<CryptoKeyPair> | undefined;
     // (undocumented)
     readonly extraHeaders: Record<string, ExtraHeader>;
     // (undocumented)
@@ -963,8 +968,6 @@ export class UserManager {
     // (undocumented)
     protected readonly _client: OidcClient;
     dpopProof(url: string, user: User, httpMethod?: string): Promise<string | undefined>;
-    // (undocumented)
-    protected readonly _dpopStore: DPoPStore | undefined;
     get events(): UserManagerEvents;
     // (undocumented)
     protected readonly _events: UserManagerEvents;
